refactor: change folder location of CarouselProfileCard component
