+++ src/frontend/src/components/Carousels/CarouselProfileCard/styles.module.scss
@@ -1,4 +1,4 @@
-@use "../../../../frontend/src/styles/variables" as v;
+@use "./../../../styles/variables" as v;
 
 .container {
     margin: 10rem;
